pyproject.toml
[project]
name = "vm-scheduler"
version = "0.1.0"
description = "Schedule libvirt VMs with systemd timers and trading-day checks."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "PyYAML>=6.0.1,<7",
  "psycopg[binary]>=3.2,<4",
]

[project.scripts]
vm-scheduler = "vm_scheduler.cli:main"

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vm_scheduler"]

[tool.uv]
package = true
